pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = 'setuptools.build_meta'

[tool.setuptools]
package-dir = {"" = "src"}

[project]
name = "neural_astar"
requires-python = ">=3.8"
dependencies = [
    "torch==2.5.1",
    "torchvision==0.20.1",
    "torchaudio==2.5.1",
    "segmentation-models-pytorch==0.3.1",
    "pqdict==1.2.0",
    "hydra-core==1.2.0",
    "numpy>=1.19.2",
    "tensorboard>=2.5",
    "moviepy>=1.0.3",
    "pytorch-lightning==1.8.5",
    "jupyterlab",
    "matplotlib",
    "tensorboardx",
    "tqdm",
    "plyfile",
    "scipy",
    "pyyaml",
    "potpourri3d",
]
version = "0.2.0"


[project.optional-dependencies]
dev = [
    "pytest",
    "pysen==0.10.1",
    "black==22.3.0",
    "flake8==4.0.1",
    "isort==5.10.1",
    "mypy==0.910",
]

[tool.pysen]
version = "0.10.1"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
py_version = "py38"

[[tool.pysen.lint.mypy_targets]]
paths = ["."]